requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.3.0
Django==3.0.3
django-jsonfield==1.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.5.0
kombu==4.6.7
pkg-resources==0.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
vine==1.3.0
zipp==3.0.0
